docs: organize sections
